Modify backend to have a functional modal with logical behaviour
Added isChef route and controller, modifield the create chefProfile route to save only userId in the user field, modified recipeMiddleWare so that it won't populate user field in every route because it was causing problem when we wanted to edit a recipe. it will save the whole user model in the user field, Also added timestamp to recipe model so that when user is fetching the recipes it will be sorted by created time, and getAllRecipesByChef was modified to fetch the correct data, before it was fetching all recipes and since recipes had wrong user field because of the malfunctional middleWare the route had weird behaviour, sometimes returning information of other chefs only.