Add cart modal and cart icon
A component is written to show the items in the cart. The component is added to our DialogControl component so we can have it as a dialog by clicking on the icon. A cart icon is added to NavBar . cartIcon is a separate component so Navbar stays as intact as possible.A cart is added to our context with related actions and reducers that supports adding to cart, removeFromCart, decreaseCount, increaseCount