Set up the backend to handle cart inside the user model
Added four routes which will support adding to cart, editing elements in cart items, deleting items in the cart, clearing the cart and getting all the information of cart. I added a field in userModel that saves the chefId and items of the card that each will be one recipeId and one qty. and upon requesting to get the cart info from backend all the fields will be populated with related information